+++ rxmicro-annotation-processor-data-sql-r2dbc-postgresql/src/test/resources/input/io.rxmicro.examples.data.r2dbc.postgresql.returntypes/delete/DeleteOneEntityFieldsUsingMonoRepository.java
@@ -34,7 +34,7 @@ public interface DeleteOneEntityFieldsUsingMonoRepository {
     Mono<Void> delete01(Long id);
 
     @Delete("DELETE FROM ${table} WHERE id=?")
-    Mono<Integer> delete02(Long id);
+    Mono<Long> delete02(Long id);
 
     @Delete("DELETE FROM ${table} WHERE id=?")
     Mono<Boolean> delete03(Long id);
