+++ rxmicro-annotation-processor-data-sql-r2dbc-postgresql/src/test/resources/input/io.rxmicro.examples.data.r2dbc.postgresql.returntypes/delete/DeleteOneEntityFieldsUsingRequiredCompletableFutureRepository.java
@@ -35,7 +35,7 @@ public interface DeleteOneEntityFieldsUsingRequiredCompletableFutureRepository {
     CompletableFuture<Void> delete01(Long id);
 
     @Delete("DELETE FROM ${table} WHERE id=?")
-    CompletableFuture<Integer> delete02(Long id);
+    CompletableFuture<Long> delete02(Long id);
 
     @Delete("DELETE FROM ${table} WHERE id=?")
     CompletableFuture<Boolean> delete03(Long id);
