+++ rxmicro-annotation-processor-data-sql-r2dbc-postgresql/src/test/resources/input/io.rxmicro.examples.data.r2dbc.postgresql.returntypes/insert/InsertOneEntityUsingMonoRepository.java
@@ -30,7 +30,7 @@ public interface InsertOneEntityUsingMonoRepository {
     Mono<Void> insert01(Account account);
 
     @Insert
-    Mono<Integer> insert02(Account account);
+    Mono<Long> insert02(Account account);
 
     @Insert
     Mono<Boolean> insert03(Account account);
