+++ rxmicro-annotation-processor-data-sql-r2dbc-postgresql/src/test/resources/input/io.rxmicro.examples.data.r2dbc.postgresql.returntypes/update/UpdateOneEntityFieldsUsingRequiredCompletionStageRepository.java
@@ -35,7 +35,7 @@ public interface UpdateOneEntityFieldsUsingRequiredCompletionStageRepository {
     CompletionStage<Void> update01(String firstName, String lastName, Long id);
 
     @Update("UPDATE ${table} SET first_name=?, last_name=? WHERE id=?")
-    CompletionStage<Integer> update02(String firstName, String lastName, Long id);
+    CompletionStage<Long> update02(String firstName, String lastName, Long id);
 
     @Update("UPDATE ${table} SET first_name=?, last_name=? WHERE id=?")
     CompletionStage<Boolean> update03(String firstName, String lastName, Long id);
