+++ rxmicro-annotation-processor-data-sql-r2dbc-postgresql/src/test/resources/input/io.rxmicro.examples.data.r2dbc.postgresql.returntypes/update/UpdateOneEntityUsingRequiredCompletionStageRepository.java
@@ -31,7 +31,7 @@ public interface UpdateOneEntityUsingRequiredCompletionStageRepository {
     CompletionStage<Void> update01(Account account);
 
     @Update
-    CompletionStage<Integer> update02(Account account);
+    CompletionStage<Long> update02(Account account);
 
     @Update
     CompletionStage<Boolean> update03(Account account);
